feat: Agregar cartas con graficas de rendimiento, calificaciones mensuales y script SQL completo - CartasScreen: Agregar grafica de tendencia historica por operario - CartasScreen: Agregar semaforo 100% con colores y porcentaje - Backend: RendimientoOperarioController para historial de operarios - Backend: RendimientoOperarioMensual modelo y DbContext actualizado - SQL: Script install_database_complete.sql consolidado - Docs: DEPLOYMENT_GUIDE.md para instalacion en servidor

+++ ProduccionUnificado/backend/DTOs/TiempoProcesoDto.cs
@@ -21,7 +21,7 @@ public class MaquinaDto
     public string Nombre { get; set; } = string.Empty;
     public int MetaRendimiento { get; set; }
     public decimal ValorPorTiro { get; set; }
-    public int Importancia { get; set; }
+    public decimal Importancia { get; set; }
     public int Meta100Porciento { get; set; }
 }
